requirements.txt
loguru==0.6.0
notifiers==1.3.3
Pillow==8.1.0
numpy==1.19.5
opencv-python==4.5.1.48
pdf2image==1.14.0
python-dateutil==2.8.1
python-dotenv==0.15.0
pytz==2020.5
pyzbar==0.1.8
setuptools==51.1.2
six==1.15.0
wheel==0.36.2



